[bphh-1797] Add better filtering of params for permit application update, disable UI elements and add more indicators for collaborators
